- Fix definition of RTL_CRITICAL_SECTION_DEBUG. - Fix all MSVC warnings in kernel32. - Re-enable PCH. - Disable Rtl*CondVar calls, since we can't target Vista *AND* 2003 due to linker problems. We'll need a way to create multiple DLL versions if we're going to attempt this. - Make ASSERTs always valid in kernel32 as long as DBG is set. - Fix Windows-loading hack, it seems there's only two levels of SEH. Eitherways, I think I've found a cleaner way to make it load nicely, so this hack will go away soon.
svn path=/trunk/; revision=25633

+++ reactos/dll/win32/kernel32/mem/global.c
@@ -914,12 +914,12 @@ GlobalMemoryStatus(LPMEMORYSTATUS lpBuffer)
         /* Reset the right size and fill out the information */
         lpBuffer->dwLength = sizeof(MEMORYSTATUS);
         lpBuffer->dwMemoryLoad = lpBufferEx.dwMemoryLoad;
-        lpBuffer->dwTotalPhys = lpBufferEx.ullTotalPhys;
-        lpBuffer->dwAvailPhys = lpBufferEx.ullAvailPhys;
-        lpBuffer->dwTotalPageFile = lpBufferEx.ullTotalPageFile;
-        lpBuffer->dwAvailPageFile = lpBufferEx.ullAvailPageFile;
-        lpBuffer->dwTotalVirtual = lpBufferEx.ullTotalVirtual;
-        lpBuffer->dwAvailVirtual = lpBufferEx.ullAvailVirtual;
+        lpBuffer->dwTotalPhys = (SIZE_T)lpBufferEx.ullTotalPhys;
+        lpBuffer->dwAvailPhys = (SIZE_T)lpBufferEx.ullAvailPhys;
+        lpBuffer->dwTotalPageFile = (SIZE_T)lpBufferEx.ullTotalPageFile;
+        lpBuffer->dwAvailPageFile = (SIZE_T)lpBufferEx.ullAvailPageFile;
+        lpBuffer->dwTotalVirtual = (SIZE_T)lpBufferEx.ullTotalVirtual;
+        lpBuffer->dwAvailVirtual = (SIZE_T)lpBufferEx.ullAvailVirtual;
     }
 }
